Add SimpleStore class.
