+++ src/Container.js
@@ -905,7 +905,7 @@ export const proxiedMapperMethods = [
  *
  * @example <caption>Container#constructor</caption>
  * // import {Container} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Container} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -957,7 +957,7 @@ export function Container (opts) {
      *
      * @example <caption>Container#mapperClass</caption>
      * // import {Container, Mapper} from 'js-data'
-     * const JSData = require('js-data@3.0.0-beta.10')
+     * const JSData = require('js-data@3.0.0-rc.4')
      * const {Container} = JSData
      * console.log('Using JSData v' + JSData.version.full)
      *
@@ -990,7 +990,7 @@ export function Container (opts) {
    *
    * @example <caption>Container#mapperDefaults</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1025,7 +1025,7 @@ const props = {
    *
    * @example <caption>Container#on</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1061,7 +1061,7 @@ const props = {
    *
    * @example <caption>Container#as</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1106,7 +1106,7 @@ const props = {
    *
    * @example <caption>Container#defineMapper</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1223,7 +1223,7 @@ const props = {
    *
    * @example <caption>Container#getMapper</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1254,7 +1254,7 @@ const props = {
    *
    * @example <caption>Container#getMapperByName</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -1319,7 +1319,7 @@ Component.extend(props)
  * Create a subclass of this Container:
  * @example <caption>Container.extend</caption>
  * // Normally you would do: import {Container} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Container} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
