+++ src/Record.js
@@ -19,6 +19,12 @@ const superMethod = function (mapper, name) {
   return mapper[name].bind(mapper)
 }
 
+// Cache these strings
+const creatingPath = 'creating'
+const noValidatePath = 'noValidate'
+const keepChangeHistoryPath = 'keepChangeHistory'
+const previousPath = 'previous'
+
 /**
  * js-data's Record class. An instance of `Record` corresponds to an in-memory
  * representation of a single row or document in a database, Firebase,
@@ -31,7 +37,7 @@ const superMethod = function (mapper, name) {
  *
  * @example <caption>Record#constructor</caption>
  * // Normally you would do: import {Record} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Record} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -45,7 +51,7 @@ const superMethod = function (mapper, name) {
  *
  * @example <caption>Record#constructor2</caption>
  * // Normally you would do: import {Mapper} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Mapper} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -59,7 +65,7 @@ const superMethod = function (mapper, name) {
  *
  * @example <caption>Record#constructor3</caption>
  * // Normally you would do: import {Container} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Container} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -72,7 +78,7 @@ const superMethod = function (mapper, name) {
  *
  * @example <caption>Record#constructor4</caption>
  * // Normally you would do: import {Container} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Container} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -91,7 +97,7 @@ const superMethod = function (mapper, name) {
  *
  * @example <caption>Record#constructor5</caption>
  * // Normally you would do: import {Container} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Container} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
@@ -123,23 +129,24 @@ function Record (props, opts) {
   props || (props = {})
   opts || (opts = {})
   const _set = this._set
-  // TODO: Optimize these strings
-  _set('creating', true)
+  _set(creatingPath, true)
   if (opts.noValidate) {
-    _set('noValidate', true)
+    _set(noValidatePath, opts.noValidate === undefined ? true : opts.noValidate)
   }
+  _set(keepChangeHistoryPath, opts.keepChangeHistory === undefined ? (mapper ? mapper.keepChangeHistory : true) : opts.keepChangeHistory)
+
   // Set the idAttribute value first, if it exists.
   const mapper = this.constructor.mapper
   const id = mapper ? utils.get(props, mapper.idAttribute) : undefined
   if (id !== undefined) {
     utils.set(this, mapper.idAttribute, id)
   }
-  const keepChangeHistory = opts.keepChangeHistory !== undefined ? opts.keepChangeHistory : (mapper ? mapper.keepChangeHistory : true)
-  _set('keepChangeHistory', keepChangeHistory)
+
   utils.fillIn(this, props)
-  _set('creating', false)
-  _set('noValidate', false)
-  _set('previous', utils.plainCopy(props))
+  _set(creatingPath, false)
+  const validateOnSet = opts.validateOnSet === undefined ? (mapper ? mapper.validateOnSet : true) : opts.validateOnSet
+  _set(noValidatePath, !validateOnSet)
+  _set(previousPath, mapper ? mapper.toJSON(props) : utils.plainCopy(props))
 }
 
 export default Component.extend({
@@ -197,7 +204,7 @@ export default Component.extend({
    *
    * @example <caption>Record#changes</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -227,7 +234,7 @@ export default Component.extend({
    *
    * @example <caption>Record#commit</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    *
@@ -283,7 +290,7 @@ export default Component.extend({
    *
    * @example <caption>Record#get</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -307,7 +314,7 @@ export default Component.extend({
    *
    * @example <caption>Record#hasChanges</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -338,7 +345,7 @@ export default Component.extend({
    *
    * @example <caption>Record#isNew</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -363,7 +370,7 @@ export default Component.extend({
    *
    * @example <caption>Record#isValid</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -552,7 +559,7 @@ export default Component.extend({
    *
    * @example <caption>Record#previous</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -585,7 +592,7 @@ export default Component.extend({
    *
    * @example <caption>Record#revert</caption>
    * // import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -690,7 +697,7 @@ export default Component.extend({
    *
    * @example <caption>Record#set</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -734,7 +741,7 @@ export default Component.extend({
    *
    * @example <caption>Record#toJSON</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -783,7 +790,7 @@ export default Component.extend({
    *
    * @example <caption>Record#unset</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -812,7 +819,7 @@ export default Component.extend({
    *
    * @example <caption>Record#validate</caption>
    * // Normally you would do: import {Container} from 'js-data'
-   * const JSData = require('js-data@3.0.0-beta.10')
+   * const JSData = require('js-data@3.0.0-rc.4')
    * const {Container} = JSData
    * console.log('Using JSData v' + JSData.version.full)
    * const store = new Container()
@@ -885,7 +892,7 @@ utils.eventify(
  * Create a subclass of this Record:
  * @example <caption>Record.extend</caption>
  * // Normally you would do: import {Record} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
+ * const JSData = require('js-data@3.0.0-rc.4')
  * const {Record} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *
